+++ lib/widgets/fields/field_base_state.dart
@@ -1,4 +1,8 @@
-import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:nylo_support/helpers/extensions.dart';
+import 'package:theme_provider/theme_provider.dart';
+import '/helpers/helper.dart';
+import '/themes/base_theme_config.dart';
 import '/widgets/ny_form.dart';
 
 import '/widgets/spacing.dart';
@@ -6,6 +10,9 @@ import '/widgets/spacing.dart';
 abstract class FieldBaseState<T extends StatefulWidget> extends State<T> {
   FieldBaseState(this.field);
   late Field field;
+
+  /// Default surface color for dark mode
+  Color surfaceColorDark = "222831".toHexColor();
 
   /// Get metadata from a field
   // ignore: avoid_shadowing_type_parameters
@@ -50,5 +57,63 @@ abstract class FieldBaseState<T extends StatefulWidget> extends State<T> {
     }
 
     return widget;
+  }
+
+  /// Get the color based on the device mode
+  Color color({Color? light, Color? dark}) {
+    bool isDarkModeEnabled = false;
+    ThemeController themeController = ThemeProvider.controllerOf(context);
+
+    if (themeController.currentThemeId == getEnv('DARK_THEME_ID')) {
+      isDarkModeEnabled = true;
+    }
+
+    if ((themeController.theme.options as NyThemeOptions).meta is Map &&
+        (themeController.theme.options as NyThemeOptions).meta['type'] ==
+            NyThemeType.dark) {
+      isDarkModeEnabled = true;
+    }
+
+    if (context.isDeviceInDarkMode) {
+      isDarkModeEnabled = true;
+    }
+
+    if (isDarkModeEnabled) {
+      return dark ?? Colors.black38;
+    }
+
+    return light ?? Colors.grey.shade100;
+  }
+
+  /// When the theme is in [light] mode, return [light] function, else return [dark] function
+  // ignore: avoid_shadowing_type_parameters
+  T whenTheme<T>({
+    required T Function() light,
+    T Function()? dark,
+  }) {
+    bool isDarkModeEnabled = false;
+    ThemeController themeController = ThemeProvider.controllerOf(context);
+
+    if (themeController.currentThemeId == getEnv('DARK_THEME_ID')) {
+      isDarkModeEnabled = true;
+    }
+
+    if ((themeController.theme.options as NyThemeOptions).meta is Map &&
+        (themeController.theme.options as NyThemeOptions).meta['type'] ==
+            NyThemeType.dark) {
+      isDarkModeEnabled = true;
+    }
+
+    if (context.isDeviceInDarkMode) {
+      isDarkModeEnabled = true;
+    }
+
+    if (isDarkModeEnabled) {
+      if (dark != null) {
+        return dark();
+      }
+    }
+
+    return light();
   }
 }