+++ lib/widgets/ny_state.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/cupertino.dart';
 import '/widgets/ny_base_state.dart';
 import '/event_bus/event_bus_plus.dart';
 import 'package:flutter/material.dart';